refactor(cli): extract common deployment utilities and update GPU defaults
* Add common.rs module with shared retry logic, env parsing, wait functions
* Consolidate WaitResult enum, create_with_retry, parse_env_vars, wait_for_ready
* Remove duplicated code from sglang.rs and vllm.rs templates
* Update default GPU from RTX A4000 (16GB) to A100 (40GB) for model sizing
* Simplify GPU recommendation to use canonical A100/H100 model names
* Update wait_for_ready to accept service_name parameter for dynamic messages
* Export common module from templates mod.rs
